+++ src/rest/middlewares/index.ts
@@ -1,4 +1,5 @@
 export * from './advancedCache.middleware';
+export * from './clearAdvancedCache.middleware';
 export * from './reportProgress.middleware';
 export * from './responseType.middleware';
 export * from './responseTransform.middleware';